requirements.txt
# this file is automatically generated using pipreqs
beautifulsoup4==4.10.0
matplotlib==3.5.1
numpy==1.22.0
opencv-python>=4.5.5
pandas==1.3.5
PyMuPDF==1.19.4
pytesseract==0.3.8
python-dotenv==0.19.2
requests==2.27.1
scikit_learn==1.0.2
seaborn==0.11.2
selenium==4.1.0
tensorflow>=2.7.0
lxml==4.7.1
